==> 0_Hageldaten_Download.ipynb <==
# 0. Download, Bereinigung und Anreicherung der zugrundeliegenden Hageldaten und Kantonsgrenzen

In diesem Notebook werden Daten zur Anzahl der Hageltage in den Sommermonaten (April-September) von 2002 bis 2022 in der Schweiz heruntergeladen und vorverarbeitet. Es handelt sich dabei um Monatsdaten pro km^2, also Gitterdaten über der Schweiz.

Da Hagel nicht flächendeckend am Boden gemessen werden können, wird ein Hageltag aus Radarmessungen abgeleitet. Ein Hageltag ist dabei definiert als ein Tag, an dem aus den Radarmessungen auf eine hohe Wahrscheinlichkeit (80%) für Hagel am Boden geschlossen wurde, basierend auf dem Radar-Parameter POH (Probability of Hail).

Weitere Informationen befinden sich in https://www.nccs.admin.ch/nccs/en/home/data-and-media-library/data/hail-climate-datasets.html.

Weiterhin werden swissBOUNDARIES3D Daten heruntergeladen. Diese beinhalten Daten zu den Grenzen der Schweiz und des Fürstentums Liechtenstein, welche später zu Visualisierungszwecken benötigt werden. Details finden sich unter https://www.swisstopo.admin.ch/de/geodata/landscape/boundaries3d.html.


Das Notebook beinhaltet dabei die folgenden Abschnitte:

* [0.1 Herunterladen der Daten](#to_download)

* [0.2 Bereinigung der Daten](#to_cleaning)

* [0.3 Anreicherung der Daten](#to_enrichment)

* [0.4 Speichern der Daten](#to_dump)

<a id="to_download"></a>
## 0.1 Herunterladen der Daten

import os
from pathlib import Path
import zipfile

#### Defintion einiger nützlicher Helferfunktionen

def unzip_data(in_file: Path, out_path: Path) -> None:
    """Unzips Zip file content into specified output path.
    
    Args:
        in_file: The zip archive to unpack.
        out_path: The output path where to unzip the Zip archive content into.
    """
    with zipfile.ZipFile(in_file, "r") as zip_ref:
        zip_ref.extractall(out_path)
        

def create_directory(directory: Path) -> None:
    """Creates a directory if it doesn't already exists.
    
    Args:
        directory: The directory to create.
    """
    directory.mkdir(parents=True, exist_ok=True)

### Herunterladen und Entpacken der Daten zur Anzahl der Hageltage von ``https://data.geo.admin.ch/`` und der swissBOUNDARIES3D Daten von ``https://www.swisstopo.admin.ch/``

# Pfad- und Dateidefinitionen für den Datendownload 
haildays_file = Path("haildays.zip")
swiss_shape_file = Path("swissboundaries3d.shp.zip")
raw_data_path = Path("./data/raw")
processed_data_path = Path("./data/processed")
shape_data_path = Path("./data/swiss_shapes")
out_hail_data_file = "haildata_per_month.csv"

# Erzeugen der Datenpfade, ein Ordner für Rohdaten und einen für prozessierte sowie für die Schweizer Shape Daten
create_directory(raw_data_path)
create_directory(processed_data_path)
create_directory(shape_data_path)

Download der Hageldaten

! wget -O {haildays_file} https://data.geo.admin.ch/ch.meteoschweiz.klima/hageltage/Swiss-hail-climatology_haildays.zip

Download der swissBOUNDARIES3D Daten

! wget -O {swiss_shape_file} https://data.geo.admin.ch/ch.swisstopo.swissboundaries3d/swissboundaries3d_2023-01/swissboundaries3d_2023-01_2056_5728.shp.zip

# Entpacken der heruntergeladenen Daten
unzip_data(haildays_file, raw_data_path)
unzip_data(swiss_shape_file, shape_data_path)

### Entfernen der zusätzlich heruntergeladenen Daten, die zur Analyse nicht benötigt werden (inklusive der Zip Archive)

# es werden nur die monatlichen Hageldaten benötigt
haildays_per_month_file = "haildaysM_ch01r.swiss.lv95_20020401000000_20220930000000.nc"
# es werden nur die Kantonsgrenzen benötigt
swiss_cantons_shapes = "swissBOUNDARIES3D_1_4_TLM_KANTONSGEBIET"

# Erzeugen zweier Listen mit allen zusätzlichen Dateien (ausser der beiden oben definierten)
files_to_delete = [
    str(x) for x in raw_data_path.iterdir()
    if x.is_file() and x.name != haildays_per_month_file
]
shape_files_to_delete = [
    str(x) for x in shape_data_path.iterdir() 
    if x.is_file() and swiss_cantons_shapes not in x.name
]

# Löschen der zusätzlich mit heruntergeladenen Daten, welche nicht für dieses Projekt benötigt werden
for file in files_to_delete:
    Path(file).unlink()
for shape_file in shape_files_to_delete:
    Path(shape_file).unlink()
    
# Entfernen der Zip Archive
haildays_file.unlink()
swiss_shape_file.unlink()

<a id="to_cleaning"></a>
## 0.2 Bereinigung der Daten

#### Installation der benötigeten Pakete

!pip install --upgrade pip

!pip install xarray
!pip install netcdf4 h5netcdf scipy pydap zarr fsspec cftime rasterio cfgrib pooch
!pip install pyaxis
!pip install numpy pandas

from pathlib import Path
import xarray as xr
from pyaxis import pyaxis
import pandas as pd
import numpy as np

#### Prozessierung der Hageldaten

haildays_ds = xr.open_dataset(raw_data_path / haildays_per_month_file)

df_hail = haildays_ds.to_dataframe()
df_hail

#### Entfernen der Datenpunlte ohne Werte für `haildays` (keine Messung vorhanden)

df_hail = df_hail[~df_hail.haildays.isna()].copy()
df_hail

#### Entfernen des `chy` und `chx` Level vom Multiindex (werden nicht benötigt, da latitude und longitude Daten vorhanden)

df_hail.index = df_hail.index.droplevel([1, 2])
df_hail

#### Konvertierung der Spalte `haildays` zu Integer

df_hail["haildays"] = df_hail["haildays"].apply(lambda x: x.days)
df_hail.head()

<a id="to_enrichment"></a>
## 0.3 Anreicherung der Daten

#### get `country`, `state` and `district` from `lon` and `lat` data via reverse geocoding

! pip install reverse_geocoder

# offline reverse geocoding
import reverse_geocoder as rg

coordinates = [(x, y) for x, y in zip(df_hail["lat"], df_hail["lon"])]
coordinates[0:10]

reverse_geocoding_results = rg.search(tuple(coordinates))

reverse_geocoding_results[0:3]

df_hail["canton"] = np.array([x["admin1"] for x in reverse_geocoding_results])
df_hail["district"] = np.array([x["admin2"] for x in reverse_geocoding_results])
df_hail["country"] = np.array([x["cc"] for x in reverse_geocoding_results])
df_hail = df_hail.reset_index()
df_hail.head()

#### Anwendung eines Kanton Mappings und Entfernen der Datenpunkte ausserhalb der Schweiz

df_hail["country"].unique()

canton_mapping = {
    "Zurich": "ZH",
    "Bern": "BE",
    "Lucerne": "LU",
    "Uri": "UR",
    "Schwyz": "SZ",
    "Obwalden": "OW",
    "Nidwalden": "NW",
    "Glarus": "GL",
    "Zug": "ZG",
    "Fribourg": "FR",
    "Solothurn": "SO",
    "Basel-City": "BS",
    "Basel-Landschaft": "BL",
    "Schaffhausen": "SH",
    "Appenzell Ausserrhoden": "AR",
    "Appenzell Innerrhoden": "AI",
    "Saint Gallen": "SG",
    "Grisons": "GR",
    "Aargau": "AG",
    "Thurgau": "TG",
    "Ticino": "TI",
    "Vaud": "VD",
    "Valais": "VS",
    "Neuchatel": "NE",
    "Geneva": "GE",
    "Jura": "JU",
}

set(df_hail["canton"].unique()) - set(canton_mapping.keys())

df_hail = df_hail[df_hail["canton"].isin(canton_mapping.keys())]
df_hail = df_hail.replace({"canton": canton_mapping})
df_hail["country"].unique()

#### Hinzufügen von separaten Spalten für Monat und Jahr

df_hail["year"] = df_hail["time"].apply(lambda x: x.year)
df_hail["month"] = df_hail["time"].apply(lambda x: x.month)

#### Entfernen von unnötigen Spalten (`swiss_coordinates`, `district`, `country`)

df_hail = df_hail.drop(["swiss_coordinates", "district", "country"], axis=1)

#### Umordnung der Spalten in schönerer Reihenfolge

columns = ["time", "year", "month", "lat", "lon", "canton", "haildays"]
df_hail = df_hail[columns]
df_hail.head()

<a id="to_dump"></a>
## 0.4 Speichern der Daten

#### Speichern der vorprozessierten Daten

df_hail.to_csv(processed_data_path / out_hail_data_file, sep=",", index=False)
