==> 1_KfZ-Zulassungsdaten_Download_Cleansing.ipynb <==
# 1. Download, Bereinigung und Vorverarbeitung der KfZ-Zulassungsdaten

Einleitung
Der Datensatz KfZ-Zulassung zeigt den jährlichen Bestand an Personenwagen in der Schweiz nach Kanton, Erstzulassungsjahr, Leergewicht, Treibstoff, Hubraum, Antriebs- und Getriebebetrieb, ab 2005.

* [1.1 Erforderliche Python-Libraries](#python_libraries)

* [1.2 KfZ-Daten herunterladen](#data_download)

* [1.3 KfZ-Daten aus Datei laden](#data_load)

* [1.4 Datenbereinigung und Vorverarbeitung](#data_cleansing)
    
* [1.5 Daten in CSV speichern](#data_saving)

#### Datenquelle: https://www.bfs.admin.ch/bfs/de/home/statistiken/kataloge-datenbanken/daten.assetdetail.24105795.html
*Daten heruntergeladen am 01.04.2023*


*Die folgenden Informationen sind der ASTRA-Website entnommen:*

<blockquote>

#### Metainformationen:  
Letzte Änderungen: neue Daten (2022), rektifizierte und revidierte Daten (2021)
Erhebungsdatum: 30. September (ausser 2022: 15. Oktober)
Raumbezug: Kantone am 1.1.2022
Erhebung: Strassenfahrzeugbestand MFZ
Datenquelle: Bundesamt für Strassen (ASTRA) - IVZ-Fahrzeuge, ehemals MOFIS-Datenbank

#### Bemerkungen: 
Die Daten zum Leergewicht und zur Getriebebedienung werden gegenwärtig revidiert. Sie sind daher vorübergehend aus dem Dataset entfernt worden. «Personenwagen» sind leichte Motorwagen zum Personentransport mit höchstens neun Sitzplätzen einschliesslich Fahrer oder Fahrerin. Der Strassenfahrzeugbestand berücksichtigt alle am 30. September in der Schweiz immatrikulierten zivilen Fahrzeuge von Haltern und Halterinnen, die an diesem Stichtag (Wohn-)Sitz in der Schweiz hatten. Aufgrund einer Bereinigung der Grunddatenbank (IVZ-Fahrzeuge) ist die Vergleichbarkeit der Daten ab 2022 mit denen der Vorjahre leicht eingeschränkt: Insgesamt rund 26'000 sich mittlerweile nicht mehr in der Schweiz befindliche Personenwagen mit Export-Kontrollschild wurden 2022 aus der Bestandesstatistik entfernt.

###### Kanton
Der Kanton wird nach der Adresse des Fahrzeughalters oder der -halterin und nicht nach dem Nummernschild bestimmt.
###### Treibstoff
*Benzin*: Bis und mit 2020 inkl. Fahrzeuge ohne Angaben zum Treibstoff (ab 2021 werden diese Fahrzeuge der Kategorie «Anderer» zugewiesen)
\n *Anderer*: Ab 2021 inkl. Fahrzeuge ohne Angaben zum Treibstoff (bis und mit 2020 waren diese Fahrzeuge der Kategorie «Benzin» zugewiesen worden)
</blockquote>

<a id="python_libraries"></a>
## 1.1 Erforderliche Python-Libraries

# Loading necessary libraries in Python
import os
from pathlib import Path
import pandas as pd
import numpy as np

# Pfad- und Dateidefinitionen für den Datendownload 
swiss_shape_file = Path("swissboundaries3d.shp.zip")
raw_data_path = Path("./data/raw")
processed_data_path = Path("./data/processed")
shape_data_path = Path("./data/swiss_shapes")


<a id="data_download"></a>
## 1.2 KfZ-Daten herunterladen von bfs.admin.ch 

Die Daten können mit dem unten stehenden Codeblock heruntergeladen werden. Da sich der Link mit der Zeit ändern kann, ist die Ausführung optional, da die Daten bereits in den Ordner data/raw/ heruntergeladen wurden.


! wget https://dam-api.bfs.admin.ch/hub/api/dam/assets/24105795/master

! mv master {raw_data_path}/px-x-1103020100_105.px

<a id="data_load"></a>
## 1.3 Laden von KfZ-Daten aus einer Datei

<div class="alert alert-block alert-info">Die Autodaten sind im .px-Format gespeichert, so dass pyaxis verwendet wird, um die Datei zu lesen und sie als Panda-Datenrahmen zu speichern.</div>

from pyaxis import pyaxis

raw_data_path = Path("./data/raw")
processed_data_path = Path("./data/processed")

kfz_file = "px-x-1103020100_105.px"

# Join the path and file name using the Path object
fp = str(raw_data_path / kfz_file)

#parse contents of *.px file
px = pyaxis.parse(uri = fp , encoding = 'ISO-8859-2')

#store data as pandas dataframe
df_kfz = px['DATA']

# In case the data is to be stored as a dictionary (not necessary in this case, but might be helpful) 
# the following line should be uncommented.  
# meta_dict = px['METADATA']

df_kfz.info()

# print the header of the dataframe
df_kfz.head()

<div class="alert alert-block alert-warning">Es sieht so aus, als ob die Spalte *Kanton* andere Werte als die 26 Kantone der Schweiz enthält.</div>

# Getting unique values in column 'Kanton'
df_kfz['Kanton'].unique()

<a id="data_cleansing"></a>
## 1.4 Datenbereinigung und Vorverarbeitung

<div class="alert alert-block alert-info">Aus diesem Grund werden alle Zeilen für "Schweiz" und "Bund" aus dem Datenrahmen gelöscht.</div> 

# Dropping all rows with string values of Schweiz and Bund in the Kanton column 
df_kfz = df_kfz[df_kfz['Kanton'].str.contains("Schweiz|Bund") == False]

df_kfz['Kanton'].unique()

<div class="alert alert-block alert-info">
Es gibt doppelte Namen für bestimmte Kantone, in Französisch und Deutsch. Für die Einheitlichkeit der Ergebnisse wird die abgekürzte Version verwendet.</div> 

# Define a dictionary with keys of German/French name and values of abbreviations. The strings in column Kanton 
# is then replaced using this dictionary. 

cantons = {
    'Zürich': 'ZH',
    'Bern / Berne': 'BE',
    'Luzern': 'LU',
    'Uri': 'UR',
    'Schwyz': 'SZ',
    'Obwalden': 'OW',
    'Nidwalden': 'NW',
    'Glarus': 'GL',
    'Zug': 'ZG',
    'Fribourg / Freiburg': 'FR',
    'Solothurn': 'SO',
    'Basel-Stadt': 'BS',
    'Basel-Landschaft': 'BL',
    'Schaffhausen': 'SH',
    'Appenzell Ausserrhoden': 'AR',
    'Appenzell Innerrhoden': 'AI',
    'St. Gallen': 'SG',
    'Graubünden / Grigioni / Grischun': 'GR',
    'Aargau': 'AG',
    'Thurgau': 'TG',
    'Ticino': 'TI',
    'Vaud': 'VD',
    'Valais / Wallis': 'VS',
    'Neuchâtel': 'NE',
    'Genčve': 'GE',
    'Jura': 'JU'
}

df_kfz = df_kfz.replace({'Kanton':cantons})

# Make a check to see if all names have been changed.
df_kfz['Kanton'].unique()

# Considering that the Jahr and DATA columns are treated as objects, following transformation are done
df_kfz.Jahr= pd.to_datetime(df_kfz.Jahr, format='%Y').dt.year
df_kfz.DATA = pd.to_numeric(df_kfz.DATA)

# Check to see if the data type has changed
df_kfz.info()

df_kfz.head()

<div class="alert alert-block alert-info">
Für die Analysen in diesem Projekt sind nur die folgenden Spalten von Bedeutung: <b>Kanton</b>, <b>Jahr</b> und <b>Daten</b>, die die Anzahl der Fahrzeuge pro Typ angibt.</div>

# Create a new dataframe with only the above columns. The data is grouped based on the "Kanton" and "Jahr" columns
# and sum of DATA which is number of car per type, is calculated.

df_kfz_grouped = df_kfz.loc[:,['Kanton', 'Jahr', 'DATA']].groupby(['Kanton','Jahr'], as_index=False)['DATA'].sum().rename(columns={'DATA':'Total_KFZ'})
#df_kfz_grouped = df_kfz.groupby(['Kanton','Jahr'])['DATA'].sum()



df_kfz_grouped.head(10)

df_kfz_grouped.info()

<a id="data_saving"></a>
## 1.5 Daten in CSV speichern

<div class="alert alert-block alert-info">
Um sicherzustellen, dass die bisherige Datenverarbeitung funktioniert hat, wird das Ergebnis in einer .csv-Datei gespeichert.</div>

df_kfz_grouped.to_csv(processed_data_path / "interim_kfz_data.csv", sep=";", index=False)
